refactor(suite-common): extract stablecoin yield state into wallet-core
