+++ packages/suite/src/components/earn/yield/common/YieldApprovedAmountCard.tsx
@@ -1,11 +1,10 @@
 import { Translation } from '@suite/intl';
 import { getCoingeckoId } from '@suite-common/wallet-config';
+import type { YieldFlowDisplayToken } from '@suite-common/wallet-core';
 import { Card, IconButton, Row, Text } from '@trezor/components';
 import { AssetLogo, CoinLogo } from '@trezor/product-components';
 
 import { FormattedCryptoAmount } from 'src/components/suite/FormattedCryptoAmount';
-
-import type { YieldFlowDisplayToken } from '../types';
 
 type YieldApprovedAmountCardProps = {
     token: YieldFlowDisplayToken;
