+++ packages/suite/src/components/earn/yield/common/YieldDisabledBanner.tsx
@@ -1,9 +1,8 @@
 import { type ReactNode } from 'react';
 
 import { Translation } from '@suite/intl';
+import type { YieldFlowType } from '@suite-common/wallet-core';
 import { Banner } from '@trezor/components';
-
-import type { YieldFlowType } from '../types';
 
 type YieldDisabledBannerProps = {
     type: YieldFlowType;
