+++ packages/suite/src/reducers/wallet/selectedAccountReducer.ts
@@ -3,12 +3,12 @@ import { type TradingState } from '@suite-common/trading';
 import {
     type AccountsRootState,
     type SendRootState,
+    type StablecoinYieldState,
     accountsActions,
     selectAccountByKey,
 } from '@suite-common/wallet-core';
 import type { SelectedAccountStatus } from '@suite-common/wallet-types';
 
-import { type YieldState } from 'src/reducers/wallet/yieldReducer';
 import type { Action } from 'src/types/suite';
 
 export type State = SelectedAccountStatus;
@@ -23,7 +23,7 @@ export type SelectedAccountRootStateWithTrading = SelectedAccountRootState &
     SendRootState & {
         wallet: {
             trading: TradingState;
-            yield: YieldState;
+            stablecoinYield: StablecoinYieldState;
         };
         connectPopup: ConnectPopupState;
     };
@@ -56,15 +56,15 @@ export const selectIsSelectedAccountLoaded = (state: SelectedAccountRootState) =
 
 /**
  * Mainly used for common modals, also used in the Trading section.
- * @returns account from yield tx review if set, then send state if set, then trading if set,
+ * @returns account from stablecoin yield tx review if set, then send state if set, then trading if set,
  * otherwise account from the store
  */
 export const selectAccountIncludingChosenInTrading = (
     state: SelectedAccountRootStateWithTrading,
 ) => {
-    const yieldAccountKey = state.wallet.yield.txReview.accountKey;
-    if (yieldAccountKey) {
-        return selectAccountByKey(state, yieldAccountKey) ?? undefined;
+    const stablecoinYieldAccountKey = state.wallet.stablecoinYield.txReview.accountKey;
+    if (stablecoinYieldAccountKey) {
+        return selectAccountByKey(state, stablecoinYieldAccountKey) ?? undefined;
     }
 
     const sendAccountKey = state.wallet.send.accountKey;
