+++ suite-common/wallet-core/src/index.ts
@@ -58,6 +58,11 @@ export * from './stake/stakeReducer';
 export type * from './stake/stakeReducerTypes';
 export * from './stake/stakeSelectors';
 export * from './stake/stakeThunks';
+export * from './stablecoinYield/stablecoinYieldReducer';
+export * from './stablecoinYield/stablecoinYieldSelectors';
+export * from './stablecoinYield/stablecoinYieldApprovalThunks';
+export * from './stablecoinYield/stablecoinYieldTypes';
+export * from './stablecoinYield/stablecoinYieldUtils';
 export * from './token/stellarTokenThunks';
 export * from './transactions/transactionsActions';
 export * from './transactions/transactionsReducer';
