+++ suite-common/wallet-core/src/stablecoinYield/stablecoinYieldTypes.ts
@@ -1,5 +1,4 @@
 import type { NetworkSymbol } from '@suite-common/wallet-config';
-import type { BulletListItemState } from '@trezor/components';
 
 export const YIELD_FLOW_TYPES = ['supply', 'withdraw'] as const;
 export const YIELD_FLOW_STEPS = ['approve', 'action', 'complete'] as const;
@@ -41,10 +40,4 @@ export type YieldPendingTransactionState = {
     type: 'approve' | 'revoke' | 'revoke-only' | 'supply' | 'withdraw';
     txid: string;
     amount: string;
-};
-
-export type UseYieldFlowStepsResult = {
-    currentStep: YieldFlowStepId;
-    stepStates: Record<YieldFlowStepId, BulletListItemState>;
-    goToStep: (step: YieldFlowStepId) => void;
 };
